+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Finished fleshing out password protected games
 + Fixed unknown error when joining a game you don't have
 + Fixed double clicking to resize columns causing a game join #810
++ Add space in username warning
 
 #3.1.42.137 - Test
 + Updated api to fix a crash